HDPI-1932: Refactor out case merging into a separate class to aid testing
