+++ src/test/java/uk/gov/hmcts/reform/pcs/ccd/service/PcsCaseServiceTest.java
@@ -5,37 +5,32 @@ import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.ArgumentCaptor;
 import org.mockito.Captor;
+import org.mockito.InOrder;
 import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
 import org.modelmapper.ModelMapper;
 import uk.gov.hmcts.ccd.sdk.type.AddressUK;
 import uk.gov.hmcts.ccd.sdk.type.ListValue;
-import uk.gov.hmcts.reform.idam.client.models.UserInfo;
 import uk.gov.hmcts.reform.pcs.ccd.domain.DefendantDetails;
 import uk.gov.hmcts.reform.pcs.ccd.domain.PCSCase;
-import uk.gov.hmcts.reform.pcs.ccd.domain.PaymentStatus;
 import uk.gov.hmcts.reform.pcs.ccd.domain.VerticalYesNo;
 import uk.gov.hmcts.reform.pcs.ccd.entity.AddressEntity;
-import uk.gov.hmcts.reform.pcs.ccd.entity.PartyEntity;
 import uk.gov.hmcts.reform.pcs.ccd.entity.PcsCaseEntity;
 import uk.gov.hmcts.reform.pcs.ccd.model.Defendant;
 import uk.gov.hmcts.reform.pcs.ccd.repository.PcsCaseRepository;
 import uk.gov.hmcts.reform.pcs.config.MapperConfig;
 import uk.gov.hmcts.reform.pcs.exception.CaseNotFoundException;
 import uk.gov.hmcts.reform.pcs.postcodecourt.model.LegislativeCountry;
-import uk.gov.hmcts.reform.pcs.security.SecurityContextService;
 
 import java.util.List;
 import java.util.Optional;
-import java.util.UUID;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.catchThrowable;
-import static org.mockito.ArgumentMatchers.any;
+import static org.mockito.Mockito.inOrder;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.spy;
 import static org.mockito.Mockito.verify;
-import static org.mockito.Mockito.verifyNoMoreInteractions;
 import static org.mockito.Mockito.when;
 
 @ExtendWith(MockitoExtension.class)
@@ -47,7 +42,7 @@ class PcsCaseServiceTest {
     @Mock
     private PcsCaseRepository pcsCaseRepository;
     @Mock
-    private SecurityContextService securityContextService;
+    private PcsCaseMergeService pcsCaseMergeService;
     @Mock
     private TenancyLicenceService tenancyLicenceService;
     @Captor
@@ -60,7 +55,7 @@ class PcsCaseServiceTest {
         MapperConfig config = new MapperConfig();
         modelMapper = spy(config.modelMapper());
 
-        underTest = new PcsCaseService(pcsCaseRepository, securityContextService, modelMapper, tenancyLicenceService);
+        underTest = new PcsCaseService(pcsCaseRepository, pcsCaseMergeService, modelMapper, tenancyLicenceService);
     }
 
     @Test
@@ -124,6 +119,36 @@ class PcsCaseServiceTest {
     }
 
     @Test
+    void shouldPatchCaseEntityAndSave() {
+        // Given
+        PCSCase pcsCase = mock(PCSCase.class);
+        PcsCaseEntity pcsCaseEntity = mock(PcsCaseEntity.class);
+        when(pcsCaseRepository.findByCaseReference(CASE_REFERENCE)).thenReturn(Optional.of(pcsCaseEntity));
+
+        // When
+        underTest.patchCase(CASE_REFERENCE, pcsCase);
+
+        // Then
+        InOrder inOrder = inOrder(pcsCaseRepository, pcsCaseMergeService);
+        inOrder.verify(pcsCaseRepository).findByCaseReference(CASE_REFERENCE);
+        inOrder.verify(pcsCaseMergeService).mergeCaseData(pcsCaseEntity, pcsCase);
+        inOrder.verify(pcsCaseRepository).save(pcsCaseEntity);
+    }
+
+    @Test
+    void shouldMergeCaseData() {
+        // Given
+        PCSCase pcsCase = mock(PCSCase.class);
+        PcsCaseEntity pcsCaseEntity = mock(PcsCaseEntity.class);
+
+        // When
+        underTest.mergeCaseData(pcsCaseEntity, pcsCase);
+
+        // Then
+        verify(pcsCaseMergeService).mergeCaseData(pcsCaseEntity, pcsCase);
+    }
+
+    @Test
     void shouldThrowExceptionPatchingUnknownCaseReference() {
         // Given
         PCSCase pcsCase = mock(PCSCase.class);
@@ -136,165 +161,6 @@ class PcsCaseServiceTest {
         assertThat(throwable)
             .isInstanceOf(CaseNotFoundException.class)
             .hasMessage("No case found with reference %s", CASE_REFERENCE);
-    }
-
-    @Test
-    void shouldLeaveFieldsUnchangedWhenPatchingCaseWithNoData() {
-        // Given
-        PCSCase pcsCase = mock(PCSCase.class);
-        PcsCaseEntity existingPcsCaseEntity = mock(PcsCaseEntity.class);
-
-        when(pcsCaseRepository.findByCaseReference(CASE_REFERENCE)).thenReturn(Optional.of(existingPcsCaseEntity));
-        when(pcsCase.getCaseManagementLocation()).thenReturn(null);
-
-        // When
-        underTest.patchCase(CASE_REFERENCE, pcsCase);
-
-        // Then
-        verify(pcsCaseRepository).save(pcsCaseEntityCaptor.capture());
-
-        PcsCaseEntity savedEntity = pcsCaseEntityCaptor.getValue();
-        assertThat(savedEntity).isSameAs(existingPcsCaseEntity);
-        verify(existingPcsCaseEntity).setTenancyLicence(any());
-        verify(existingPcsCaseEntity).setPossessionGrounds(any());
-        verifyNoMoreInteractions(existingPcsCaseEntity);
-    }
-
-    @Test
-    void shouldChangeFieldsWhenPatchingCase() {
-        // Given
-        PCSCase pcsCase = mock(PCSCase.class);
-
-        AddressUK updatedPropertyAddress = mock(AddressUK.class);
-        final AddressEntity updatedAddressEntity = stubAddressUKModelMapper(updatedPropertyAddress);
-
-        PcsCaseEntity existingPcsCaseEntity = mock(PcsCaseEntity.class);
-        when(pcsCase.getPropertyAddress()).thenReturn(updatedPropertyAddress);
-
-        when(pcsCaseRepository.findByCaseReference(CASE_REFERENCE)).thenReturn(Optional.of(existingPcsCaseEntity));
-
-        // When
-        underTest.patchCase(CASE_REFERENCE, pcsCase);
-
-        // Then
-        verify(pcsCaseRepository).save(pcsCaseEntityCaptor.capture());
-
-        PcsCaseEntity savedEntity = pcsCaseEntityCaptor.getValue();
-        assertThat(savedEntity).isSameAs(existingPcsCaseEntity);
-        verify(existingPcsCaseEntity).setPropertyAddress(updatedAddressEntity);
-    }
-
-    @Test
-    void shouldCreatePartyWithPcqIdForCurrentUser() {
-        UUID userId = UUID.randomUUID();
-        UUID expectedPcqId = UUID.randomUUID();
-        String expectedForename = "some forename";
-        String expectedSurname = "some surname";
-
-        PCSCase pcsCase = mock(PCSCase.class);
-        when(pcsCase.getUserPcqId()).thenReturn(expectedPcqId.toString());
-
-        UserInfo userDetails = mock(UserInfo.class);
-        when(userDetails.getUid()).thenReturn(userId.toString());
-        when(userDetails.getGivenName()).thenReturn(expectedForename);
-        when(userDetails.getFamilyName()).thenReturn(expectedSurname);
-        when(securityContextService.getCurrentUserDetails()).thenReturn(userDetails);
-
-        PcsCaseEntity existingPcsCaseEntity = new PcsCaseEntity();
-
-        when(pcsCaseRepository.findByCaseReference(CASE_REFERENCE)).thenReturn(Optional.of(existingPcsCaseEntity));
-
-        // When
-        underTest.patchCase(CASE_REFERENCE, pcsCase);
-
-        // Then
-        verify(pcsCaseRepository).save(pcsCaseEntityCaptor.capture());
-        PcsCaseEntity savedEntity = pcsCaseEntityCaptor.getValue();
-        assertThat(savedEntity.getParties())
-            .hasSize(1)
-            .allSatisfy(party -> {
-                assertThat(party.getIdamId()).isEqualTo(userId);
-                assertThat(party.getPcqId()).isEqualTo(expectedPcqId);
-                assertThat(party.getForename()).isEqualTo(expectedForename);
-                assertThat(party.getSurname()).isEqualTo(expectedSurname);
-            });
-
-    }
-
-    @Test
-    void shouldUpdateExistingPartyWithPcqIdForCurrentUser() {
-        final UUID userId = UUID.randomUUID();
-        final UUID expectedPcqId = UUID.randomUUID();
-        final UUID existingPartyId = UUID.randomUUID();
-
-        PCSCase pcsCase = mock(PCSCase.class);
-        when(pcsCase.getUserPcqId()).thenReturn(expectedPcqId.toString());
-
-        UserInfo userDetails = mock(UserInfo.class);
-        when(userDetails.getUid()).thenReturn(userId.toString());
-        when(securityContextService.getCurrentUserDetails()).thenReturn(userDetails);
-
-        PcsCaseEntity existingPcsCaseEntity = new PcsCaseEntity();
-        PartyEntity existingPartyEntity = new PartyEntity();
-        existingPartyEntity.setId(existingPartyId);
-        existingPartyEntity.setIdamId(userId);
-        existingPcsCaseEntity.addParty(existingPartyEntity);
-
-        when(pcsCaseRepository.findByCaseReference(CASE_REFERENCE)).thenReturn(Optional.of(existingPcsCaseEntity));
-
-        // When
-        underTest.patchCase(CASE_REFERENCE, pcsCase);
-
-        // Then
-        verify(pcsCaseRepository).save(pcsCaseEntityCaptor.capture());
-        PcsCaseEntity savedEntity = pcsCaseEntityCaptor.getValue();
-        assertThat(savedEntity.getParties())
-            .hasSize(1)
-            .allSatisfy(
-                party -> {
-                    assertThat(party.getId()).isEqualTo(existingPartyId);
-                    assertThat(party.getPcqId()).isEqualTo(expectedPcqId);
-                }
-            );
-
-    }
-
-    @Test
-    void shouldUpdatePaymentStatusWhenNotNull() {
-        // Given
-        PCSCase pcsCase = mock(PCSCase.class);
-        PcsCaseEntity existingPcsCaseEntity = mock(PcsCaseEntity.class);
-        PaymentStatus paymentStatus = PaymentStatus.PAID;
-
-        when(pcsCase.getPaymentStatus()).thenReturn(paymentStatus);
-        when(pcsCaseRepository.findByCaseReference(CASE_REFERENCE)).thenReturn(Optional.of(existingPcsCaseEntity));
-
-        // When
-        underTest.patchCase(CASE_REFERENCE, pcsCase);
-
-        // Then
-        verify(pcsCaseRepository).save(pcsCaseEntityCaptor.capture());
-        verify(existingPcsCaseEntity).setPaymentStatus(paymentStatus);
-        assertThat(pcsCaseEntityCaptor.getValue()).isSameAs(existingPcsCaseEntity);
-    }
-
-    @Test
-    void shouldUpdateCaseManagementLocationWhenNotNull() {
-        // Given
-        PCSCase pcsCase = mock(PCSCase.class);
-        PcsCaseEntity existingPcsCaseEntity = mock(PcsCaseEntity.class);
-        Integer location = 13685;
-
-        when(pcsCase.getCaseManagementLocation()).thenReturn(location);
-        when(pcsCaseRepository.findByCaseReference(CASE_REFERENCE)).thenReturn(Optional.of(existingPcsCaseEntity));
-
-        // When
-        underTest.patchCase(CASE_REFERENCE, pcsCase);
-
-        // Then
-        verify(pcsCaseRepository).save(pcsCaseEntityCaptor.capture());
-        verify(existingPcsCaseEntity).setCaseManagementLocation(location);
-        assertThat(pcsCaseEntityCaptor.getValue()).isSameAs(existingPcsCaseEntity);
     }
 
     @Test
@@ -334,25 +200,6 @@ class PcsCaseServiceTest {
 
         // Then
         verify(pcsCaseRepository).save(pcsCaseEntity);
-    }
-
-    @Test
-    void shouldUpdatePreActionProtocolCompletedWhenNotNull() {
-        // Given
-        PCSCase pcsCase = mock(PCSCase.class);
-        PcsCaseEntity existingPcsCaseEntity = mock(PcsCaseEntity.class);
-        VerticalYesNo preActionProtocolCompleted = VerticalYesNo.YES;
-
-        when(pcsCase.getPreActionProtocolCompleted()).thenReturn(preActionProtocolCompleted);
-        when(pcsCaseRepository.findByCaseReference(CASE_REFERENCE)).thenReturn(Optional.of(existingPcsCaseEntity));
-
-        // When
-        underTest.patchCase(CASE_REFERENCE, pcsCase);
-
-        // Then
-        verify(pcsCaseRepository).save(pcsCaseEntityCaptor.capture());
-        verify(existingPcsCaseEntity).setPreActionProtocolCompleted(preActionProtocolCompleted.toBoolean());
-        assertThat(pcsCaseEntityCaptor.getValue()).isSameAs(existingPcsCaseEntity);
     }
 
     @Test
